requirements.txt
flask
flask_restful
uuid
mysql_connector_python